@@ -1,7 +1,7 @@
 
 source 'http://rubygems.org'
 ruby '2.3.1'
-gem 'rails', '4.2.7.1'
+gem 'rails', '5.0.0.1'
 gem 'mysql2'
 
 gem 'puma'
@@ -25,6 +25,7 @@ source 'https://rails-assets.org' do
 end
 
 gem 'auto_html', '2.0.0'
+gem 'aws-sdk', '< 2'
 gem 'cancancan'
 
 gem 'calendrier'
@@ -57,7 +58,7 @@ gem 'haml'
 gem 'has_scope'
 gem 'hominid'#, '2.1.9'
 gem 'http_accept_language'
-gem 'inherited_resources'
+# gem 'inherited_resources'
 gem 'instagram'
 gem 'jquery-infinite-pages'
 gem 'jquery-rails'
@@ -77,14 +78,14 @@ gem 'meta-tags', :require => 'meta_tags'
 gem 'momentjs-rails'
 gem 'nested_form' #, :git => 'git://github.com/ryanb/nested_form.git'
 
-gem 'paperclip'#, "~> 3.0"
+gem 'paperclip', :git=> 'https://github.com/thoughtbot/paperclip', :ref => '523bd46c768226893f23889079a7aa9c73b57d68'
 gem 'public_activity'
 gem 'rack-cache'
 gem "rack-cors", group: :development
 gem 'rails-timeago', '~> 2.0'
 gem 'randumb'
 gem 'recaptcha', :require => 'recaptcha/rails'
-gem 'responders'
+gem 'responders', '>= 2.3.0'
 gem 'redcarpet'
 gem 'rolify'
 gem 'sass'
