@@ -5,8 +5,8 @@ class Project < ActiveRecord::Base
   has_many :events
   belongs_to :location
   has_attached_file :avatar, :styles => {:new_carousel => "960x400#", :medium => "400x400>", :thumb => "100x100>" },
-          :path =>  ":rails_root/public/images/projects/:id/:style/:basename.:extension", 
-        :url => "/images/projects/:id/:style/:basename.:extension", :default_url => "/assets/missing.png"
+               url: ':s3_domain_url',
+        path: "/projects/:id/:style/:basename.:extension", :default_url => "/assets/missing.png"
   has_attached_file :carousel, :styles => {:largest => "1180x492#",
                                           :new_carousel => "960x400#", :full => "600x400#", :small => "300x200#", :thumb => "100x100>"}, :path =>  ":rails_root/public/images/carousel/projects/:id/:style/:basename.:extension",
                           :url =>  "/images/carousel/projects/:id/:style/:basename.:extension"
