@@ -4,7 +4,7 @@ class Admin::BaseController < ApplicationController
   respond_to :html
   #load_and_authorize_resource
   # check_authorization
-  skip_before_filter :require_no_authentication
+  # skip_before_filter :require_no_authentication
 
   def check_permissions
     authorize! :create, resource
