@@ -2,7 +2,7 @@
 require File.expand_path('../boot', __FILE__)
 
 require 'rails/all'
-
+require 'i18n/backend/fallbacks'
 if defined?(Bundler)
   # If you precompile assets before deploying to production, use this line
   Bundler.require(*Rails.groups(:assets => %w(development test)))
@@ -55,6 +55,7 @@ module Ptarmigan
     # Configure the default encoding used in templates for Ruby 1.9.
     config.encoding = "utf-8"
     config.i18n.fallbacks = true
+    
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
     config.autoload_paths += %W(#{config.root}/app/models/ckeditor)
@@ -63,7 +64,7 @@ module Ptarmigan
     config.active_support.escape_html_entities_in_json = true
     config.assets.precompile += %w( *.woff *.eot *.svg *.ttf)
     config.assets.precompile += %w( common.css print.css  jwplayer.js jwplayer.html5.js cms.css cms.js creativeterritories/javascripts/application.js creativeterritories/javascripts/staff.js madhouse/stylesheets/staff.css creativeterritories/stylesheets/application.css madhouse/javascripts/staff.js )
-    config.i18n.fallbacks = true
+
     # Use SQL instead of Active Record's schema dumper when creating the database.
     # This is necessary if your schema can't be completely dumped by the schema dumper,
     # like if you have constraints or database-specific column types
